set %OnNew call based on constructor args.
This commit fixes an issue on which when calling the %OnNew (constructor) implicitly, that would cause to lose the track of events.

This behavior has been changed to only call it implicitly when asked to do so.

In the case of mocks and stubs, it's triggered if both receive constructor arguments.